rename global to link
