+++ README.md
@@ -16,17 +16,16 @@ This would install and run [SwiftLint](https://github.com/realm/SwiftLint) versi
 Mint is designed to be used with Swift command line tools that build with the Swift Package Manager. It makes installing, running and distributing these tools much easier.
 
 - ✅ easily run a specific **version** of a package
-- ✅ install a package **globally**
+- ✅ **link** a package **globally**
 - ✅ builds are **cached** by version
 - ✅ use **different versions** of a package side by side
 - ✅ easily run the **latest** version of a package
 - ✅ distribute your own packages **without recipes and formulas**
 - ✅ specify a list of versioned packages in a **Mintfile** for easy use
 
-Homebrew is a popular method of distributing Swift executables, but that requires creating a formula and then maintaining that formula. Running specific versions of homebrew installations can also be tricky as only one global version is installed at any one time. Mint installs your package via SPM and lets you run multiple versions of that package, which are globally installed and cached on demand.
+Homebrew is a popular method of distributing Swift executables, but that requires creating a formula and then maintaining that formula. Running specific versions of homebrew installations can also be tricky as only one global version is installed at any one time. Mint installs your package via SPM and lets you run multiple versions of that package, which are installed and cached in a central place.
 
 If your Swift executable package builds with SPM, then it can be run with Mint! See [Support](#support) for details.
-
 
 ## Why is it called Mint?
 Swift Packager Manager Tools -> SPMT -> Spearmint -> Mint! 🌱😄
@@ -101,15 +100,15 @@ An optional version can be specified by appending `@version`, otherwise the newe
 #### Examples
 ```sh
 $ mint run yonaskolb/XcodeGen@1.2.4 xcodegen --spec spec.yml # pass some arguments
-$ mint install yonaskolb/XcodeGen@1.2.4 --prevent-global # installs a certain version but not globally
+$ mint install yonaskolb/XcodeGen@1.2.4 --no-link # installs a certain version but doesn't link it globally
 $ mint install yonaskolb/XcodeGen # install newest tag
 $ mint install yonaskolb/XcodeGen@master --force #reinstall the master branch
 $ mint run yonaskolb/XcodeGen@1.2.4 # run 1.2.4
 $ mint run XcodeGen # use newest tag and find XcodeGen in installed packages
 ```
 
-### Global installs
-By default Mint symlinks your installs into `usr/local/bin` when `install` or `update` are used, unless `--prevent-global` is passed. This means a package will be accessible from anywhere, and you don't have to prepend commands with `mint run package`. Note that only one global version can be installed at a time though. If you need to run a specific older version use `mint run`.
+### Linking
+By default Mint symlinks your installs into `usr/local/bin` on `mint install`, unless `--no-link` is passed. This means a package will be accessible from anywhere, and you don't have to prepend commands with `mint run package`. Note that only one linked version can be used at a time though. If you need to run a specific older version use `mint run`.
 
 ### Mintfile
 A `Mintfile` can specify a list of versioned packages. It makes installing and running these packages easy, as the specific repos and versions are centralized.
@@ -135,7 +134,7 @@ mint bootstrap
 
 ### Advanced
 - You can use `--silent` in `mint run` to silence any output from mint itself. Useful if forwarding output somewhere else.
-- You can set `MINT_PATH` and `MINT_INSTALL_PATH` envs to configure where mint caches builds, and where it symlinks global installs. These default to `/usr/local/lib/mint` and `/usr/local/bin` respectively
+- You can set `MINT_PATH` and `MINT_LINK_PATH` envs to configure where mint caches builds, and where it symlinks global installs. These default to `/usr/local/lib/mint` and `/usr/local/bin` respectively
 - You can use `mint install --force` to reinstall a package even if it's already installed. This shouldn't be required unless you are pointing at a branch and want to update it.
 
 ### Linux
