+++ Sources/MintCLI/Commands/ListCommand.swift
@@ -6,7 +6,7 @@ class ListCommand: MintCommand {
     init(mint: Mint) {
         super.init(mint: mint,
                    name: "list",
-                   description: "List all the currently installed packages. Globally installed packages are marked with *")
+                   description: "List all the currently installed packages. Globally linked packages are marked with *")
     }
 
     override func execute() throws {
