+++ Sources/MintCLI/MintCLI.swift
@@ -14,16 +14,16 @@ public class MintCLI {
     public init() {
 
         var mintPath: Path = "/usr/local/lib/mint"
-        var installationPath: Path = "/usr/local/bin"
+        var linkPath: Path = "/usr/local/bin"
 
         if let path = ProcessInfo.processInfo.environment["MINT_PATH"], !path.isEmpty {
             mintPath = Path(path)
         }
-        if let path = ProcessInfo.processInfo.environment["MINT_INSTALL_PATH"], !path.isEmpty {
-            installationPath = Path(path)
+        if let path = ProcessInfo.processInfo.environment["MINT_LINK_PATH"], !path.isEmpty {
+            linkPath = Path(path)
         }
 
-        mint = Mint(path: mintPath, installationPath: installationPath)
+        mint = Mint(path: mintPath, linkPath: linkPath)
 
         cli = CLI(name: "mint", version: version, description: "Run and install Swift Package Manager executables", commands: [
             RunCommand(mint: mint),
